Move clock related functions to Clock in clock.cc.
This CL factors out clock module from base:util to sort out
dependencies.

BUG=
TEST=unittest
REF_BUG=19010851
REF_CL=84935258
