+++ src/base/util.cc
@@ -1037,161 +1037,6 @@ void Util::SetRandomSeed(uint32 seed) {
   ::srand(seed);
 }
 
-namespace {
-class ClockImpl : public Util::ClockInterface {
- public:
-#ifndef __native_client__
-  ClockImpl() {}
-#else  // __native_client__
-  ClockImpl() : timezone_offset_sec_(0) {}
-#endif  // __native_client__
-  virtual ~ClockImpl() {}
-
-  virtual void GetTimeOfDay(uint64 *sec, uint32 *usec) {
-#ifdef OS_WIN
-    FILETIME file_time;
-    GetSystemTimeAsFileTime(&file_time);
-    ULARGE_INTEGER time_value;
-    time_value.HighPart = file_time.dwHighDateTime;
-    time_value.LowPart = file_time.dwLowDateTime;
-    // Convert into microseconds
-    time_value.QuadPart /= 10;
-    // kDeltaEpochInMicroSecs is difference between January 1, 1970 and
-    // January 1, 1601 in microsecond.
-    // This number is calculated as follows.
-    // ((1970 - 1601) * 365 + 89) * 24 * 60 * 60 * 1000000
-    // 89 is the number of leap years between 1970 and 1601.
-    const uint64 kDeltaEpochInMicroSecs = 11644473600000000ULL;
-    // Convert file time to unix epoch
-    time_value.QuadPart -= kDeltaEpochInMicroSecs;
-    *sec = static_cast<uint64>(time_value.QuadPart / 1000000UL);
-    *usec = static_cast<uint32>(time_value.QuadPart % 1000000UL);
-#else  // OS_WIN
-    struct timeval tv;
-    gettimeofday(&tv, NULL);
-    *sec = tv.tv_sec;
-    *usec = tv.tv_usec;
-#endif  // OS_WIN
-  }
-
-  virtual uint64 GetTime() {
-#ifdef OS_WIN
-    return static_cast<uint64>(_time64(NULL));
-#else
-    return static_cast<uint64>(time(NULL));
-#endif  // OS_WIN
-  }
-
-  virtual bool GetTmWithOffsetSecond(time_t offset_sec, tm *output) {
-    const time_t current_sec = static_cast<time_t>(this->GetTime());
-    const time_t modified_sec = current_sec + offset_sec;
-
-#ifdef OS_WIN
-    if (_localtime64_s(output, &modified_sec) != 0) {
-      return false;
-    }
-#elif defined(__native_client__)
-    const time_t localtime_sec = modified_sec + timezone_offset_sec_;
-    if (gmtime_r(&localtime_sec, output) == NULL) {
-      return false;
-    }
-#else  // !OS_WIN && !__native_client__
-    if (localtime_r(&modified_sec, output) == NULL) {
-      return false;
-    }
-#endif  // OS_WIN
-    return true;
-  }
-
-  virtual uint64 GetFrequency() {
-#if defined(OS_WIN)
-    LARGE_INTEGER timestamp;
-    // TODO(yukawa): Consider the case where QueryPerformanceCounter is not
-    // available.
-    const BOOL result = ::QueryPerformanceFrequency(&timestamp);
-    return static_cast<uint64>(timestamp.QuadPart);
-#elif defined(OS_MACOSX)
-    static mach_timebase_info_data_t timebase_info;
-    mach_timebase_info(&timebase_info);
-    return static_cast<uint64>(
-        1.0e9 * timebase_info.denom / timebase_info.numer);
-#elif defined(OS_LINUX)
-    return 1000000uLL;
-#else  // platforms (OS_WIN, OS_MACOSX, OS_LINUX, ...)
-#error "Not supported platform"
-#endif  // platforms (OS_WIN, OS_MACOSX, OS_LINUX, ...)
-  }
-
-  virtual uint64 GetTicks() {
-    // TODO(team): Use functions in <chrono> once the use of it is approved.
-#if defined(OS_WIN)
-    LARGE_INTEGER timestamp;
-    // TODO(yukawa): Consider the case where QueryPerformanceCounter is not
-    // available.
-    const BOOL result = ::QueryPerformanceCounter(&timestamp);
-    return static_cast<uint64>(timestamp.QuadPart);
-#elif defined(OS_MACOSX)
-    return static_cast<uint64>(mach_absolute_time());
-#elif defined(OS_LINUX)
-    uint64 sec;
-    uint32 usec;
-    GetTimeOfDay(&sec, &usec);
-    return sec * 1000000 + usec;
-#else  // platforms (OS_WIN, OS_MACOSX, OS_LINUX, ...)
-#error "Not supported platform"
-#endif  // platforms (OS_WIN, OS_MACOSX, OS_LINUX, ...)
-  }
-
-#ifdef __native_client__
-  virtual void SetTimezoneOffset(int32 timezone_offset_sec) {
-    timezone_offset_sec_ = timezone_offset_sec;
-  }
-
- private:
-  int32 timezone_offset_sec_;
-#endif  // __native_client__
-};
-
-Util::ClockInterface *g_clock_handler = NULL;
-
-Util::ClockInterface *GetClockHandler() {
-  if (g_clock_handler != NULL) {
-    return g_clock_handler;
-  } else {
-    return Singleton<ClockImpl>::get();
-  }
-}
-
-}  // namespace
-
-void Util::SetClockHandler(Util::ClockInterface *handler) {
-  g_clock_handler = handler;
-}
-
-void Util::GetTimeOfDay(uint64 *sec, uint32 *usec) {
-  GetClockHandler()->GetTimeOfDay(sec, usec);
-}
-
-uint64 Util::GetTime() {
-  return GetClockHandler()->GetTime();
-}
-
-bool Util::GetCurrentTm(tm *current_time) {
-  return GetTmWithOffsetSecond(current_time, 0);
-}
-
-bool Util::GetTmWithOffsetSecond(tm *time_with_offset, int offset_sec) {
-  return GetClockHandler()->GetTmWithOffsetSecond(offset_sec, time_with_offset);
-}
-
-uint64 Util::GetFrequency() {
-  return GetClockHandler()->GetFrequency();
-}
-
-uint64 Util::GetTicks() {
-  return GetClockHandler()->GetTicks();
-}
-
 void Util::Sleep(uint32 msec) {
 #ifdef OS_WIN
   ::Sleep(msec);
@@ -1199,12 +1044,6 @@ void Util::Sleep(uint32 msec) {
   usleep(msec * 1000);
 #endif  // OS_WIN
 }
-
-#ifdef __native_client__
-void Util::SetTimezoneOffset(int32 timezone_offset_sec) {
-  return GetClockHandler()->SetTimezoneOffset(timezone_offset_sec);
-}
-#endif  // __native_client__
 
 namespace {
