+++ src/dictionary/file/codec.cc
@@ -32,6 +32,7 @@
 #include <algorithm>
 #include <climits>
 
+#include "base/clock.h"
 #include "base/file_stream.h"
 #include "base/hash.h"
 #include "base/logging.h"
