+++ src/rewriter/date_rewriter_test.cc
@@ -31,6 +31,7 @@
 
 #include <cstddef>
 
+#include "base/clock.h"
 #include "base/clock_mock.h"
 #include "base/port.h"
 #include "base/scoped_ptr.h"
@@ -222,7 +223,7 @@ class DateRewriterTest : public testing::Test {
 TEST_F(DateRewriterTest, DateRewriteTest) {
   scoped_ptr<ClockMock> mock_clock(
       new ClockMock(kTestSeconds, kTestMicroSeconds));
-  Util::SetClockHandler(mock_clock.get());
+  Clock::SetClockForUnitTest(mock_clock.get());
 
   DateRewriter rewriter;
   Segments segments;
@@ -471,7 +472,7 @@ TEST_F(DateRewriterTest, DateRewriteTest) {
     }
   }
 
-  Util::SetClockHandler(NULL);
+  Clock::SetClockForUnitTest(nullptr);
 }
 
 TEST_F(DateRewriterTest, ADToERA) {
